backend/requirements.txt
fastapi
uvicorn
gunicorn
wandb
pydantic
python-dotenv
pandas